+++ czkawka_core/src/bad_extensions.rs
@@ -231,9 +231,7 @@ impl BadExtensions {
 
                 true
             }
-            DirTraversalResult::SuccessFolders { .. } => {
-                unreachable!()
-            }
+
             DirTraversalResult::Stopped => false,
         }
     }
