+++ czkawka_core/src/empty_files.rs
@@ -77,9 +77,7 @@ impl EmptyFiles {
 
                 true
             }
-            DirTraversalResult::SuccessFolders { .. } => {
-                unreachable!()
-            }
+
             DirTraversalResult::Stopped => false,
         }
     }
