+++ krokiet/src/connect_scan.rs
@@ -3,11 +3,12 @@ use crate::{CurrentTab, GuiState, MainListModel, MainWindow, ProgressToSend};
 use chrono::NaiveDateTime;
 use crossbeam_channel::{Receiver, Sender};
 use czkawka_core::common::{split_path, split_path_compare, DEFAULT_THREAD_SIZE};
-use czkawka_core::common_dir_traversal::{FileEntry, FolderEntry, ProgressData};
+use czkawka_core::common_dir_traversal::{FileEntry, ProgressData};
 use czkawka_core::common_tool::CommonData;
 use czkawka_core::common_traits::ResultEntry;
 use czkawka_core::empty_files::EmptyFiles;
 use czkawka_core::empty_folder::EmptyFolder;
+use czkawka_core::empty_folder::FolderEntry;
 use czkawka_core::similar_images;
 use czkawka_core::similar_images::SimilarImages;
 use humansize::{format_size, BINARY};
